dev: use dev and test DBs without changing env
Previously we would always set the full database URL (including the
database name) from the `DATABASE_URL` env var on startup, requiring
that it be provided alongside each command such as `mix phx.server` or
`mix test`. This presents an obstacle to using Direnv or shell exports,
since the value needs to change frequently depending on the command.

With this change, in dev and test the app expects a `DATABASE_ROOT_URL`
that is a Postgres URL minus the database name, and appends the correct
name for the environment. Outside dev and test, `DATABASE_URL` is still
used to set up the repo at runtime.