feat: add multi-select and groups to stop filter
The stop filter is now a `<select multiple>` and has the conventional
behavior for submitting multiple values for the same form field, i.e.
the URL is `filters[stop_ids][]=70001&filters[stop_ids][]=...` rather
than `filters[stop_id]=70001`.

Our existing `bootstrap-combobox` library didn't support multi-select,
so this swaps it out for Selectize. Unfortunately this library is not
actively maintained (last release in 2006), but it is otherwise stable
and one of the few that doesn't require React/Vue/etc., works correctly
out of the box with our Webpack/jQuery/Bootstrap setup, and has search
ranking (i.e. typing "Wonderland" puts the station itself at the top,
above all the stops that have "Wonderland" elsewhere in their name and
would be sorted above it alphabetically).

To avoid creating very long URLs, stop groups are implemented as single
"special values" (such as `_terminal`) on the client side, which the
server translates into a set of stop IDs. For now these are hard-coded.