fix: VehiclePositions.Tracker prefers argument for AWS URL over env
With the addition of direnv to the application
get_env_vehicle_positions_url is now generally non-null when you run
`mix test` at the command line locally. This was causing it to
override the value passed in by some tests. Instead, we should prefer
the value given in the arguments when starting the tracker over the
default environment variable value.