perf: adjust database indexes for usage patterns
This makes the following index changes:

* Drop `trip_id` on `vehicle_events`, since this index appears unused
  (both in the code and as indicated by Postgres stats), and is taking
  up ~11GB of space that could be reclaimed.

* Add `stop_id` on `predictions`, since this column is queried in the
  process of associating vehicle events with predictions. The index is
  partial on rows where `vehicle_event_id IS NULL`, since this condition
  is always included in the queries that include `stop_id`. This should
  save a bit of disk space since the condition is only true for half the
  rows in the current production database.

* Replace `vehicle_id` on `predictions` with a partial index having the
  same condition as the above, for the same reason.

* Add `stop_id` and `vehicle_id` on `vehicle_events`, partial on rows
  where `departure_time IS NULL`, since these columns are queried when
  filling in departure times.

* Add `stop_id` on `prediction_accuracy`, since filtering on this field
  is part of the frontend's regular operation.

Also considered but *not* implemented were indexes on `predictions` for
`route_id`, and on `prediction_accuracy` for `route_id` and `bin`. While
we do query on these fields often, it's always in combination with at
least one other indexed field; and since there are only a few distinct
routes and bins, the index likely wouldn't narrow the result set enough
to justify its use (to the query planner, if not to us).