feat: track and filter by kind of prediction
RTR predictions internally have a "kind" which is:

* "at terminal" when the train is waiting to depart from a terminal
* "reverse" when the prediction is for a trip we expect will occur once
  the train reverses direction at the end of its current trip
* "mid-trip" in all other circumstances

These are not exposed to consumers, but they are mapped to specific
values of the `uncertainty` field in stop time updates. Here we reverse
that mapping and store/aggregate the "kind" of predictions. This can be
null, since the concept is not relevant to Commuter Rail predictions.