Node 18 (#698)
* chore: upgrade to Node 18.15.0

Switches from node-sass to sass and sass-loader as node-sass is
deprecated and incompatible with Node 18. Upgrades webpack from v4 to
v5 as webpack v4 is incompatible with Node 18. Switches from
optimize-css-assets-webpack-plugin to css-minimizer-webpack-plugin as
optimize-css-assets-webpack-plugin is incompatible with webpack
v5. Upgrades from TypeScript v3 to v5 as v3 was broken by other
upgrades.

* fix: remove npm test step from CI as there aren't any